requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==2.10.0
async-generator==1.10
attrs==23.1.0
behave==1.2.6
certifi==2023.5.7
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
PySocks==1.7.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
wsproto==1.2.0
